Restructure to hexagonal/onion architecture
