+++ graphqshell/Cargo.toml
@@ -19,7 +19,7 @@ anyhow = "1.0"
 
 directories = "3.0"
 rand = "0.8.4"
-clap = "3.0.0-beta.2"
+clap = "3.0.0-beta.5"
 serde = "1.0"
 
 # config
