+++ graphqshell/src/adapter/configuration.rs
