+++ graphqshell/src/adapter/mod.rs
@@ -1,0 +1,3 @@
+pub mod configuration;
+pub mod graphql_api;
+pub mod tui_app;
