+++ graphqshell/src/adapter/tui_app/tui/activity.rs
