+++ graphqshell/src/adapter/tui_app/tui/components/log.rs
