+++ graphqshell/src/adapter/tui_app/tui/context.rs
