--- graphqshell/src/frontend.rs
@@ -1,1 +1,0 @@
-pub mod tui;
